Refactor AI server action code
