+++ src/auth/index.ts
@@ -44,7 +44,7 @@ export const {
   },
 });
 
-export const safelyRunServerAdminAction = async <T>(
+export const safelyRunAdminServerAction = async <T>(
   callback: () => T,
 ): Promise<T> => {
   const session = await auth();
