+++ src/photo/ai.ts
@@ -1,0 +1,13 @@
+import { streamOpenAiImageQuery } from '@/services/openai';
+
+export type ImageQuery = 'title' | 'caption' | 'tags' | 'description';
+
+export const IMAGE_QUERIES: Record<ImageQuery, string> = {
+  title: 'What is the title of this image?',
+  caption: 'What is the caption of this image?',
+  tags: 'Describe this image three or less comma-separated keywords',
+  description: 'Describe this image in detail',
+};
+
+export const streamImageQuery = (imageBase64: string, query: ImageQuery) =>
+  streamOpenAiImageQuery(imageBase64, IMAGE_QUERIES[query]);
