+++ src/services/openai.ts
@@ -5,7 +5,10 @@ import { createStreamableValue, render } from 'ai/rsc';
 
 const provider = new OpenAI({ apiKey: process.env.OPENAI_SECRET_KEY });
 
-const streamImageQueryRaw = async (imageBase64: string, query: string) => {
+export const streamOpenAiImageQuery = async (
+  imageBase64: string,
+  query: string,
+) => {
   const stream = createStreamableValue('');
 
   render({
@@ -36,15 +39,3 @@ const streamImageQueryRaw = async (imageBase64: string, query: string) => {
 
   return stream.value;
 };
-
-export type ImageQuery = 'title' | 'caption' | 'tags' | 'description';
-
-export const IMAGE_QUERIES: Record<ImageQuery, string> = {
-  title: 'What is the title of this image?',
-  caption: 'What is the caption of this image?',
-  tags: 'Describe this image three or less comma-separated keywords',
-  description: 'Describe this image in detail',
-};
-
-export const streamImageQuery = (imageBase64: string, query: ImageQuery) =>
-  streamImageQueryRaw(imageBase64, IMAGE_QUERIES[query]);
